==> src/cleaning/1_dataset_aggregation.ipynb <==
from pathlib import Path
import pandas as pd
import glob
import re

import warnings
warnings.filterwarnings('ignore')

from IPython.display import display
pd.options.display.max_columns = None

HOME = Path().resolve().parent.parent
data_dir = str(HOME)
input_dir = data_dir + '/data/backup/'
output_dir = data_dir + '/data/processed/'

### 1. Read all the files into a single dataset

all_files = glob.glob(input_dir + "*.csv.gz")
li = []
regex = re.compile(r'\d+')

print('FILEDATE \tMIN_DATE \tMAX_DATE \tDATASET_SHAPE')

for filename in all_files:
    
    # Read the raw data
    df = pd.read_csv(filename, index_col=None)
    
    # Save filename date for future analysis   
    file_date = re.search(regex, filename).group(0)
    df['file_date'] = file_date
    
    if ('CasosCovMes' not in df.columns.values) and ('CasosCovMesAnt' not in df.columns.values):
        df['CasosCovMes'] = 0
        df['CasosCovMesAnt'] = 0
    
    # Print data info for each file
    print('{} \t\t {} \t {} \t {}'.format(file_date, df.last_scraped.min(), df.last_scraped.max(), df.shape))
    
    # Append the data in a list
    li.append(df)

We must pay attention in the datasets from March, 2020 to September, 2020 (7 datasets). Each of then contains a sample with inconsistent columns that must be removed. 

The are some trash in the ```last_scraped``` column and it must be cleaned later. Also, we are going to save the file date info for future analysis. 

valid_columns = set(li[0].columns.values)

for i, df in enumerate(li):    
    current_cols = set(df.columns.values)
    valid_columns.intersection_update(current_cols)
    
print('There are {} valid columns in the dataset.'.format(len(valid_columns)))

for i in range(len(li)):
    li[i] = li[i][valid_columns]

# Aggregate all the data into a sigle dataset:
df = pd.concat(li, axis=0, ignore_index=True)

### 2. Removing columns pointed by Pol:

# Removed 'bathrooms_text' and 'number_of_reviews_l30d' from cols2drop

cols2drop = ['listing_url', 'scrape_id', 'name', 'description', 'neighborhood_overview', 
             'picture_url', 'host_id', 'host_url', 'host_name', 'host_about', 
             'host_response_time', 'host_response_rate', 'host_acceptance_rate', 
             'host_is_superhost', 'host_thumbnail_url', 'host_picture_url', 
             'host_listings_count', 'host_total_listings_count', 'host_verifications',
             'host_has_profile_pic', 'host_identity_verified', 
             'minimum_nights', 'maximum_nights', 'host_location',
             'minimum_minimum_nights', 'maximum_minimum_nights', 'minimum_maximum_nights', 
             'maximum_maximum_nights', 'minimum_nights_avg_ntm', 'maximum_nights_avg_ntm', 
             'host_since', 'neighbourhood', 'calendar_updated', 'review_scores_communication', 
             'review_scores_location', 'license', 'calculated_host_listings_count',
             'calculated_host_listings_count_entire_homes', 
             'calculated_host_listings_count_private_rooms', 
             'calculated_host_listings_count_shared_rooms', 'host_neighbourhood', 
             'neighbourhood_cleansed', 'review_scores_accuracy', 'review_scores_cleanliness', 
             'review_scores_checkin', 'review_scores_value', 'number_of_reviews_ltm', 
             'first_review', 'last_review', 'instant_bookable']

listings = df.drop(columns=cols2drop).copy()
rows, cols = listings.shape
print('The current dataset has {} rows and {} columns.'.format(rows, cols))

rows, cols = listings.shape
print('The final dataset has {} rows and {} columns.'.format(rows, cols))

### Save the final, aggregated dataset:

listings.to_csv(output_dir + '/listings_agg.csv.gz', index=False)
