==> src/cleaning/2_missing_values_removal.ipynb <==
from pathlib import Path
import pandas as pd

import warnings
warnings.filterwarnings('ignore')

from IPython.display import display
pd.options.display.max_columns = None

%matplotlib inline

HOME = Path().resolve().parent.parent
data_dir = str(HOME) + '/data/processed/'

df = pd.read_csv(data_dir + 'listings_agg.csv.gz')

rows, cols = df.shape
print('The original dataset has {} rows and {} columns.'.format(rows, cols))

listings = df.copy()

## Dealing with missing values

count_na = df.isna().sum()
count_na[count_na > 0].sort_values(ascending=False)

### 1. Remove the rows where there are less that 1% of missing values

These 7 samples with missing values are those we have seen during the aggreagation process. Thus, removing then now, we do not have to deal with it during the cleaning process.

listings['id'] = listings['id'].map(lambda x: pd.to_numeric(x, errors='coerce'))
listings['review_scores_rating'] = listings['review_scores_rating'].map(lambda x: pd.to_numeric(x, errors='coerce'))

# If a column has less than 1% of missing values and it is not trivial filling it, the row can be removed:
listings = listings.dropna(
    subset=['latitude', 'price' , 
          'number_of_reviews', 'longitude', 'availability_365', 
          'availability_60', 'file_date', 'has_availability', 'id', 
          'availability_30', 'neighbourhood_group_cleansed', 
          'availability_90', 'calendar_last_scraped', 'last_scraped']).copy()

It count how many remaining missing values are in the dataset:

count_na = listings.isna().sum()
count_na[count_na > 0].sort_values(ascending=False)

The most of ```bathrooms```, ```bedrooms``` and ```beds``` are filled by 1. So:

listings[['bathrooms', 'bedrooms', 'beds']] = listings[['bathrooms', 'bedrooms', 'beds']].fillna(1)

For the ```reviews_per_month``` and ```reviews_per_month```, we need a different approach:
 + Fill in with zero when there is no reviews at all
 + Fill in with the median of the reviews for the same listing

def fill_reviews(df, identif, review):
    
    # Case 1: Take the id of places without any review and set its review to zero
    count_reviews = df[[identif, review]].groupby(identif).count()
    ids_no_reviews = count_reviews[count_reviews[review] == 0].index
    idx1 = df[df[identif].isin(ids_no_reviews)].index
    df.loc[idx1, review] = 0.0

    # Case 2: If a place has some 'review' and some missing value, 
    # fill in the missing value with the median of its review
    median_review = df[[identif, review]].groupby(identif).median()
    idx2 = df[df[review].isna()].index
    df.loc[idx2, review] = df.loc[idx2, identif].map(lambda x: median_review.loc[x])
    
    return df

Filling in the missing values in the ```review_scores_rating``` column:

listings = fill_reviews(listings, 'id', 'review_scores_rating')

Filling the missing values in the ```reviews_per_month``` column:

listings = fill_reviews(listings, 'id', 'reviews_per_month')

The dataset now has no missing values

count_na = listings.isna().sum()
count_na[count_na > 0].sort_values(ascending=False)

### 4. Save processed dataset

listings.to_csv(data_dir + 'listings_notna.csv.gz', index=False)
