==> src/cleaning/3_cleaning_listings.ipynb <==
from pathlib import Path
import pandas as pd
import re

import warnings
warnings.filterwarnings('ignore')

from IPython.display import display
pd.options.display.max_columns = None

HOME = Path().resolve().parent.parent
data_dir = str(HOME) + '/data/processed/'

df = pd.read_csv(data_dir + 'listings_notna.csv.gz', parse_dates=['last_scraped'])
df.shape

listings = df.copy()

listings.head()

### 1. Dropping useless columns

cols2drop = ['has_availability', 'calendar_last_scraped']

# All values of 'has_availability' == 't'
# The date in 'calendar_last_scraped' is equal to 'last_scraped'
listings = listings.drop(columns=cols2drop)

### 2. Dealing with numerical columns

numerical_features = ['beds','bedrooms', 'number_of_reviews', 'price', 'longitude', 'latitude', 
                      'bathrooms', 'id', 'reviews_per_month', 'availability_30', 'availability_60', 
                     'availability_90', 'availability_365', 'accommodates', 'review_scores_rating']

float2int = ['beds','bedrooms', 'number_of_reviews', 'bathrooms', 'id', 'availability_30', 
             'availability_60', 'availability_90', 'availability_365', 'accommodates']

listings[float2int] = listings[float2int].applymap(lambda x: int(x))

Remove any symbol different than digit and '.' from numerical float columns:

ncols2clean = ['price', 'reviews_per_month', 'review_scores_rating']

regex = re.compile(r'(\d+\.\d{1,2})')
listings[ncols2clean] = listings[ncols2clean].applymap(lambda x: pd.to_numeric(re.search(regex, x).group(0)))

### 3. Removing listings with no avaibility

idx2drop = listings[
    (listings['availability_30'] == 0) & 
    (listings['availability_60'] == 0) &
    (listings['availability_90'] == 0) & 
    (listings['availability_365'] == 0)].index

print('Removing {} rows from the dataset.'.format(len(idx2drop)))

unavailable_listings = listings.loc[idx2drop]
listings = listings.drop(idx2drop)

### 5. Save clean data

listings.to_csv(data_dir + 'listings_clean.csv.gz', index=False)

unavailable_listings.to_csv(data_dir + 'unavailable_listings.csv.gz', index=False)
